Make every format string agree with its arguments
Fifty-eight printf-style calls disagreed with what they were passed, and
nothing could see them. Format is a plain vsprintf and Error a plain
vsnprintf, both varargs, and the build compiles with -w.

Twenty-six passed arguments the string never consumed. Three passed FEWER
than it consumed, so vsprintf printed whatever the register file held next --
those were done first, because they are the ones that can crash. Three were
invalid specifiers, and two of those three were one bug: ASSERT pasted the
asserted expression INTO its format string, so any % in the expression became
a conversion specifier. That is fixed at the macro, so every future ASSERT is
immune. Twenty-six were the wrong type for their specifier.

Nine of them put one of the engine's own <Res> tags in a string that neither
Format nor Error expands, and passed the argument anyway. Those printed the
tag literally and dropped the value. They are real defects, not false
positives against the engine's syntax, and there were no false positives at
all.

Thirty-eight of the fifty-eight are claimed as upstream's and marked as one
class rather than fifty-eight times over. The other twenty are not claimed:
fourteen are upstream code merely non-portable to LP64, and six are ours,
made by this port narrowing int32 and hObj from long to int under format
strings that still said %ld.

The compiler is the oracle and getting it to speak needed a change of its
own. -w is not an ordinary warning switch that a later flag overrides; it
suppresses the diagnostic engine outright, so EXTRA_CXXFLAGS could never
revive a warning. build_macos.sh gained a WARN_FLAGS knob that REPLACES it
and defaults to -w, leaving every ordinary build byte-identical.

Fifty-eight warnings before, zero after. The attribute stays on Format, and
tools/check_format_strings.sh fails if the count ever rises above the
baseline file beside it -- proved red by deleting one argument.

Behaviour is unchanged where it can be measured: dive seed 3 under the pinned
settings still reports 1972 turns and no errors.

+++ src/Art.cpp
@@ -328,7 +328,7 @@ PRIVATE int print_item(int p)
       k++;
    }
    if (yygrammar[i] <= 0) printf(" *");
-   T1->CWrite(Format(" (back:%d sub:%d left:%d)\n", back[p], sub[p], left[p]));
+   T1->CWrite(Format(" (back:%d sub:%d left:%d)\n", (int)back[p], (int)sub[p], (int)left[p]));
 }
 
 /*----------------------------------------------------------------------------*/
