+++ src/Tokens.cpp
@@ -1064,7 +1064,7 @@ do_action:	/* This label is used only to access EOF actions. */
 case 1:
 YY_RULE_SETUP
 #line 235 "lang\\Tokens.Lex"
-{ sscanf(yytext,"#line \"%s %d",preprocFilename,&yypos);
+{ sscanf(yytext,"#line \"%s %ld",preprocFilename,&yypos);
                       yypos --; /* we'll read the \n next */
                       for (int16 i=0;preprocFilename[i] && i != 1022;i++)
                         {
@@ -1126,14 +1126,14 @@ YY_RULE_SETUP
                else if (szl && sz != szl && !warned)
                  {
                    warned = true;
-                   yyerror(Format("Error: Mismatched grid line at line %d.",ln));
+                   yyerror(Format("Error: Mismatched grid line at line %d.",(int)ln));
                  }
                szl = 0;
              }
                 
            }
          if (ch1 == EOF)
-           { yyerror(Format("Error: Unterminated Map Grid starting at line %d.",yypos));
+           { yyerror(Format("Error: Unterminated Map Grid starting at line %d.",(int)yypos));
              return ARCHERY; }
          if ((ch2 = yyinput()) == '}')
            {
@@ -2612,7 +2612,7 @@ void CatLiteral(const char*str)
                           LBuff[lpos++] = '>'; break;
                 case 0: break;
                 default:
-                  yyerror(Format("Unknown escape sequence at line %d.", yypos));
+                  yyerror(Format("Unknown escape sequence at line %d.", (int)yypos));
               }
           else {      
             if (i)
